Add files via upload
This GoogleDrive3.0 project is a decentralized application that uses IPFS services to store and retrieve files. These files are published by users and can be accessed by other users with correct address. This method ensures the privacy of the files, as they can only be accessed by people who have been given the address.